server/pyproject.toml
[tool.poetry]
name = "serverzebra"
version = "0.1.0"
description = ""
authors = ["giannifer7 <gianni.ferrarotti@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0.0"
asgiref = "^3.5.2"
click = "^8.1.3"
colorama = "^0.4.4"
fastapi = "^0.78.0"
greenlet = "^1.1.2"
h11 = "^0.13.0"
pydantic = "^1.9.0"
SQLAlchemy = "^1.4.36"
sqlalchemy2-stubs = "^0.0.2-alpha.22"
sqlmodel = "^0.0.6"
typing-extensions = "^4.2.0"
uvicorn = "^0.17.6"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
requests = "^2.27.1"
nox = "^2022.1.7"
fastapi = "^0.78.0"
flake8-bugbear = "^22.4.25"
black = "^22.3.0"
mypy = "^0.960"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]
